get BCad.Core.UI to build

+++ BCad.Core.UI/BCadDialog.xaml.cs
@@ -1,17 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace BCad.UI
 {
@@ -62,7 +52,7 @@ namespace BCad.UI
 
         private void Window_KeyDown(object sender, KeyEventArgs e)
         {
-            if (e.Key == Key.Escape)
+            if ((int)e.Key == (int)Key.Escape)
             {
                 Cancel();
             }
